Field descriptors are now classes that instantiate field values.
This is similar to the descriptor idea, but instead of implementing 
`__get__`, the FieldContainer class constructor creates instances of all 
member descriptors and rebinds the attribute names. This means that, for 
example, `VRTPrologue.stream_id` returns the class while 
`prologue.stream_id` returns an instance. This way, the classes (field 
descriptors) hold the static information, such as the name, while 
instances hold data that varies packet-to-packet, like the value and 
enablement.