Handle packet timestamp parsing differently.
Packet timestamps are now configured with a "timestamp" entry which can only contain "integer" and "fractional" keys, whose values are just the TSI/TSF codes. There is no separate configuration for TSI/TSF and the timestamp fields.

As part of the changes, the generic mapping parser that supports registering sub-fields has been factored out of the section parser, which can also fetch packet fields as part of field parsing.