Add Spectrum fields in CIF1 data model
Currently, if you define a simple context packet that uses the spectrum
field in CIF1, vrtpktgen throws an in the code generation phase. For
example, this YAML:

```yaml
ExampleContext: !Context
  ...
  cif_1: !CIF1
    spectrum: optional
```

...will produce:

```
vrtpktgen cpp ./example-context.yaml
<--- cpp generation stack trace --->
jinja2.exceptions.UndefinedError: 'None' has no attribute 'reserved_bits'
```

The CIF1 data model appears to have a placeholder for the Spectrum field
defined in the VITA 49.2 standard in section 9.6.1.

Add the data structures described in section 9.6.1 to avoid this error
and use the spectrum field in CIF1.

Signed-off-by: John Moon <john.moon@vts-i.com>