Fix endianess and size bug in spectrum fields
Currently, some subfields in the spectrum field are not being swapped
correctly which leads to inverted values on the wire. Additionally,
compiler padding causes additional padded data to be sent.

To fix these issues, add swapping where necessary and set the size to
the static value of 13 words (per the VITA 49.2 spec).